remove output directory adjustments
It is not our business to care about symlinks ("latest") or prefix
with a hash / timestamp / name. Keep it simple by having rebuild
taking (a) the old directory (to rebuild) and (b) the output directory
for the rebuild.

The build subcommand takes the output directory (which can be relative to
cwd!).